Switched from codyco-superbuild to robotology-superbuild
I took advantage of the new superbuild to add to the image the upcoming WB-Toolbox version and the new WholeBodyControllers repository. They will be included soon in the robotology-superbuild.

Many environment variables now are handled by the superbuild installation itself and have been removed from the dockerfile.

Furthermore, some dirty stuff: in yarp-matlab-bindings and codyco-modules, the devel branch must be checked out in order to compile smoothly. It is still somehow a dirty operation on superbuilds (https://github.com/robotology/robotology-superbuild/issues/33)