Create a systemwide python virtualenv
This is my docker-based solution to share python packages (installed with sudo pip during the build stage) with the runtime user.

The trick is to create a system-wide and transparent virtualenv with access to system packages, and configuring the environment to have the resulting env constantly active.
In this way, pip called by the user will not download again in its local folder packages that are installed by root. This behavior is sound only because this is a docker image, and the packages installed in the image should match the requirement of the project.